code/requirements.txt
numpy==1.17.3
pandas==0.25.3
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
tensorflow==2.2.0
keras==2.3.1
scikit-learn==0.23.1
tqdm==4.46.0
nltk==3.5